Move pylxd.operation to pylxd.models.operation
