+++ pylxd/models/__init__.py
@@ -2,3 +2,4 @@ from pylxd.models.certificate import Certificate  # NOQA
 from pylxd.models.container import Container, Snapshot  # NOQA
 from pylxd.models.image import Image  # NOQA
 from pylxd.models.network import Network  # NOQA
+from pylxd.models.operation import Operation  # NOQA
