+++ pylxd/models/image.py
@@ -14,7 +14,7 @@
 import hashlib
 
 from pylxd.models import _model as model
-from pylxd.operation import Operation
+from pylxd.models.operation import Operation
 
 
 def _image_create_from_config(client, config, wait=False):
