+++ pylxd/models/operation.py
